requirements.txt
asgiref==3.5.2
autopep8==1.7.0
boto3==1.24.60
botocore==1.27.60
Django==4.1
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
